backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
Werkzeug==3.0.1
xgboost==2.0.3
lightgbm==4.1.0

